+++ src/request.rs
@@ -1,18 +1,7 @@
+use crate::partial::Partial;
 use async_trait::async_trait;
 use axum::extract::FromRequestParts;
 use http::{request::Parts, HeaderMap, HeaderValue, StatusCode};
-
-/// Partial reload data.
-///
-/// Clients can request a subset of the props if a page component is
-/// being refreshed. They must also include a desired component -- the
-/// server may respond with a different end component, which will
-/// include a full response.
-#[derive(Clone, Debug)]
-pub(crate) struct Partial {
-    pub(crate) props: Vec<String>,
-    pub(crate) component: String,
-}
 
 /// Inertia-related information in the request.
 ///